gitee-pages-action/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
decorator==5.1.0
idna==3.3
py==1.11.0
pyasn1==0.4.8
PyYAML==6.0
requests==2.26.0
retry==0.9.2
rsa==4.8
urllib3==1.26.7
